Segmentation (#37)
* compiles using cmake; we'll use cmake in setup.py later

* setup.py works

* move CMakeList.txt file to cpp folder

* cmake build output file setup

* test skeleton fixed

* make segmentation a separate module
