+++ mac-dev-environment.yml
@@ -1,0 +1,73 @@
+name: reneu
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - boost-cpp=1.73.0=he5d75e3_3
+  - bzip2=1.0.8=haf1e3a3_3
+  - ca-certificates=2020.6.20=hecda079_0
+  - certifi=2020.6.20=py38h32f6830_0
+  - gtest=1.8.0=h0a44026_4
+  - icu=67.1=hb1e8313_0
+  - libblas=3.8.0=17_openblas
+  - libcblas=3.8.0=17_openblas
+  - libcxx=10.0.1=h5f48129_0
+  - libedit=3.1.20191231=h1de35cc_1
+  - libffi=3.3=hb1e8313_2
+  - libgfortran=4.0.0=3
+  - libgfortran4=7.5.0=h1565451_3
+  - liblapack=3.8.0=17_openblas
+  - libopenblas=0.3.10=openmp_h63d9170_4
+  - llvm-openmp=10.0.1=h28b9765_0
+  - lz4-c=1.9.2=hb1e8313_3
+  - ncurses=6.2=h0a44026_1
+  - numpy=1.19.1=py38h8ccc501_2
+  - openssl=1.1.1g=haf1e3a3_1
+  - pip=20.2.2=py38_0
+  - python=3.8.5=h26836e1_0
+  - python_abi=3.8=1_cp38
+  - readline=8.0=h1de35cc_0
+  - setuptools=49.6.0=py38_0
+  - sqlite=3.33.0=hffcf06c_0
+  - tbb=2019.9=ha1b3eb9_1
+  - tk=8.6.10=hb0a8c7a_0
+  - wheel=0.34.2=py38_0
+  - xsimd=7.4.8=ha1b3eb9_0
+  - xtensor=0.21.5=ha1b3eb9_0
+  - xtensor-blas=0.17.2=h5a8e21a_0
+  - xtensor-python=0.24.1=py38hbf1eeb5_0
+  - xtl=0.6.18=h879752b_0
+  - xz=5.2.5=h1de35cc_0
+  - zlib=1.2.11=h1de35cc_3
+  - zstd=1.4.5=h289c70a_2
+  - pip:
+    - bleach==3.1.5
+    - chardet==3.0.4
+    - colorama==0.4.3
+    - docutils==0.16
+    - flake8==3.8.3
+    - greenlet==0.4.16
+    - idna==2.10
+    - keyring==21.4.0
+    - mccabe==0.6.1
+    - msgpack==1.0.0
+    - neovim==0.3.1
+    - packaging==20.4
+    - pkginfo==1.5.0.1
+    - ptvsd==4.3.2
+    - pybind11==2.5.0
+    - pycodestyle==2.6.0
+    - pyflakes==2.2.0
+    - pygments==2.6.1
+    - pynvim==0.4.1
+    - pyparsing==2.4.7
+    - readme-renderer==26.0
+    - requests==2.24.0
+    - requests-toolbelt==0.9.1
+    - rfc3986==1.4.0
+    - six==1.15.0
+    - tqdm==4.48.2
+    - twine==3.2.0
+    - urllib3==1.25.10
+    - webencodings==0.5.1
+prefix: /Users/jwu/opt/anaconda3/envs/reneu
