+++ mac-environment.yml
@@ -1,4 +1,3 @@
-name: reneu
 channels:
   - conda-forge
   - defaults
@@ -67,4 +66,3 @@ dependencies:
     - twine==3.2.0
     - urllib3==1.25.10
     - webencodings==0.5.1
-prefix: /Users/jwu/opt/anaconda3/envs/reneu
