+++ tests/test_nblast.py
@@ -8,9 +8,9 @@ from math import isclose
 from copy import deepcopy
 from time import time
 
-from reneu.libreneu import XNBLASTScoreTable
+from reneu.lib.skeleton import XNBLASTScoreTable
+from reneu.lib.skeleton import XVectorCloud, XNBLASTScoreMatrix
 from reneu.skeleton import Skeleton
-from reneu.libreneu import XVectorCloud, XNBLASTScoreMatrix
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), '../data/')
 #DATA_DIR = 'data/'
