+++ tests/test_segmentation.py
@@ -1,4 +1,4 @@
-from reneu.libreneu import XSupervoxelDendrogram as SupervoxelDendrogram
+from reneu.lib.segmentation import XSupervoxelDendrogram as SupervoxelDendrogram
 import h5py
 import os
 import tifffile
@@ -6,16 +6,16 @@ import numpy as np
 
 def test_supervoxel_dendrogram():
     DIR = os.path.join(os.path.dirname(__file__), '../data/')
-    with h5py.File(os.path.join(DIR, "affs_test.h5"), "r") as f:
+    with h5py.File(os.path.join(DIR, "aff_160k.h5"), "r") as f:
         affs = np.asarray(f["main"])
     
     print('read fragments from watershed...')
-    with h5py.File(os.path.join(DIR, "fragments_test.h5"), "r") as f:
-        fragments = np.asarray(f["main"][:, :300, :300])
+    with h5py.File(os.path.join(DIR, "ffn_seg.h5"), "r") as f:
+        fragments = np.asarray(f["seg"])
      
     print('construct dendrogram and merge supervoxels')
-    dend = SupervoxelDendrogram(affs, fragments, 0.3)
-    seg = dend.segment(0.3)
+    dend = SupervoxelDendrogram(affs, fragments, 0.7)
+    seg = dend.segment(0.7)
     
     print('save results...')
     tifffile.imwrite(os.path.join(DIR, "seg_test.tif"), data=seg)
